\name{cleanInvalid-package}
\alias{cleanInvalid-package}
\alias{cleanInvalid}
\docType{package}
\title{
\packageTitle{cleanInvalid}
}
\description{
\packageDescription{cleanInvalid}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cleanInvalid}
\packageIndices{cleanInvalid}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cleanInvalid}

Maintainer: \packageMaintainer{cleanInvalid}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
