% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_empty.na.nan.null.inf.R
\name{replace_empty.na.nan.null.inf}
\alias{replace_empty.na.nan.null.inf}
\title{Replace occurances of any empty, NA, NaN, NULL, or Inf  found in data}
\usage{
replace_empty.na.nan.null.inf(x, replace = 0, ...)
}
\arguments{
\item{replace_*}{convienence wrappers for various replace invalid numeric data}
}
\value{
x
}
\description{
Replace occurances of any empty, NA, NaN, NULL, or Inf  found in data
}
\examples{
message(replace_negative((-50:49),0))
test.err.a <-c(-1/0,-49:48,-Inf)
test.err.b <-c(sqrt(-12),-49:48,"")
test.all <- replace_empty.na.nan.null.inf(rbind(test.err.a,test.err.b),0)

}
\keyword{replace}
