% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_empty.R
\name{replace_empty}
\alias{replace_empty}
\title{Replace characters of 0 length}
\usage{
replace_empty(x, replace = 0, ...)
}
\arguments{
\item{x}{data to be searched for characters of 0 length that should be replaced}

\item{replace}{the data to replace with}
}
\value{
x data with all NA values replaced
}
\description{
Replace characters of 0 length
}
\examples{
replace_null(NULL,-49:48,NA)
}
\keyword{replace}
