% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_infinite.R
\name{replace_infinite}
\alias{replace_infinite}
\title{Replace infinate values}
\usage{
replace_infinite(x, replace = 0, ...)
}
\arguments{
\item{x}{data to be searched for Inf values that should be replaced}

\item{replace}{the data to replace with}
}
\value{
x data with all Inf values replaced
}
\description{
Replace infinate values
}
\examples{
replace_nan(-Inf,-49:48,Inf)
}
