% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na.nan.neg.null.inf.R
\name{replace_na.nan.neg.null.inf}
\alias{replace_na.nan.neg.null.inf}
\title{Replace occurances of any NA,NaN,Negative,NULL, or Inf  found in data}
\usage{
replace_na.nan.neg.null.inf(x, replace = 0, ...)
}
\arguments{
\item{replace_*}{convienence wrappers for various replace invalid numeric data}
}
\value{
x
}
\description{
Replace occurances of any NA,NaN,Negative,NULL, or Inf  found in data
}
\examples{
test.err.a <-c(-1/0,-49:48,-Inf)
test.err.b <-c(sqrt(-12),-49:49)
test.all <- replace_na.nan.neg.null.inf(rbind(test.err.a,test.err.b),0)
}
\keyword{replace}
