% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na.R
\name{replace_na}
\alias{replace_na}
\title{Replace NA values with}
\usage{
replace_na(x, replace = 0, ...)
}
\arguments{
\item{x}{data to be searched for NA values that should be replaced}

\item{replace}{the data to replace with}
}
\value{
x data with all NA values replaced
}
\description{
Replace NA values with
}
\examples{
replace_null(NULL,-49:48,NA)
}
