% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_negative.R
\name{replace_negative}
\alias{replace_negative}
\title{Replace negative values}
\usage{
replace_negative(x, replace = 0, ...)
}
\arguments{
\item{x}{data to be searched for negative values that should be replaced}

\item{replace}{the data to replace with}
}
\value{
x data with all negative values replaced
}
\description{
Replace negative values
}
\examples{
replace_negative((-50:49),0)
replace_negative(matrix(-50:49,10,10),0)
}
