% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_null.R
\name{replace_null}
\alias{replace_null}
\title{Replace NULL values}
\usage{
replace_null(x, replace = 0, ...)
}
\arguments{
\item{x}{data to be searched for NULL values that should be replaced}

\item{replace}{the data to replace with}
}
\value{
x data with all NULL values replaced
}
\description{
Replace NULL values
}
\examples{
replace_null(NULL,-49:48,Inf)
}
